bug fixes re threading, race conditions, deadlocks, etc (see newbugs.md under this commit for details)

+++ recorder/include/recorder/common/data_types.h
@@ -4,7 +4,9 @@
 #include <opencv2/opencv.hpp>
 
 struct FrameData {
-    unsigned int frame_id = -1;
+    // Matches the acquirers' frame counters (signed long); -1 until the frame is
+    // assigned an id during recording.
+    long frame_id = -1;
     uint64_t acquisition_time = 0; // as returned by frame grabber
     uint64_t received_time = 0;    // as returned by frame grabber
     cv::Mat image;
